add api logic with src and split
